fixed a bug/point of confusion where the *row* indexes for Excel spreadsheets were shorts. these were modified to be type int, since otherwise problems occur in error reporting and log output (and potentially other places) when a worksheet has more than 2^15-1 (~32K) rows (as is the case with Human4 duplex library)
git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1257 80c535d0-b110-0410-b4de-9279f94142c3

+++ src/edu/harvard/med/screensaver/io/libraries/rnai/RNAiLibraryColumnHeaders.java
@@ -134,12 +134,12 @@ class RNAiLibraryColumnHeaders
    * @param rowIndex the index of the data row in the sheet
    * @return a map from required columns to the contents of the row for that column
    */
-  Map<RequiredRNAiLibraryColumn,String> getDataRowContents(HSSFRow dataRow, short rowIndex)
+  Map<RequiredRNAiLibraryColumn,String> getDataRowContents(HSSFRow dataRow, int rowIndex)
   {
     Map<RequiredRNAiLibraryColumn,String> dataRowContents =
       new HashMap<RequiredRNAiLibraryColumn,String>();
     for (RequiredRNAiLibraryColumn column : RequiredRNAiLibraryColumn.values()) {
-      Cell cell = _cellFactory.getCell(getColumnIndex(column), (short) rowIndex);
+      Cell cell = _cellFactory.getCell(getColumnIndex(column), rowIndex);
       dataRowContents.put(column, cell.getAsString());      
     }
     return dataRowContents;
@@ -170,7 +170,7 @@ class RNAiLibraryColumnHeaders
   private boolean populateColumnIndexes() {
     boolean hasNoDuplicateHeaders = true;
     for (short i = _columnHeaderRow.getFirstCellNum(); i <= _columnHeaderRow.getLastCellNum(); i++) {
-      Cell cell = _cellFactory.getCell(i, (short) 0);
+      Cell cell = _cellFactory.getCell(i, 0);
       String columnHeader = cell.getString();
       if (columnHeader == null) {
         continue;
